vmw_pvscsi: Expand vcpuHint to 16 bit to aliagn with host side change. [no ci]
vcpuHint has been expanded to 16 bit on host side to enable
interruptions to be routed to more CPUs. Guest side should align with
the change.

This change has been tested with hosts with 8-bit and 16-bit vcpuHint,
on both platforms host side can get correct value.

This driver is for ESXi product which only supports x86/x64. They are
little-endian. So there is no need to consider big-endian system.

PR:		264840
Reviewed by:	imp@, Zhenlei Huang